Raw handling: remove duplicate code
