+++ packages/blocks/src/api/raw-handling/utils.js
@@ -12,21 +12,18 @@ import { isPhrasingContent } from '@wordpress/dom';
  * Internal dependencies
  */
 import { hasBlockSupport } from '..';
+import { getRawTransforms } from './get-raw-transforms';
 
 /**
  * Given raw transforms from blocks, merges all schemas into one.
  *
- * @param {Array}  transforms            Block transforms, of the `raw` type.
  * @param {Object} phrasingContentSchema The phrasing content schema.
  * @param {Object} isPaste               Whether the context is pasting or not.
  *
  * @return {Object} A complete block content schema.
  */
-export function getBlockContentSchema(
-	transforms,
-	phrasingContentSchema,
-	isPaste
-) {
+export function getBlockContentSchema( phrasingContentSchema, isPaste ) {
+	const transforms = getRawTransforms();
 	const schemas = transforms.map( ( { isMatch, blockName, schema } ) => {
 		const hasAnchorSupport = hasBlockSupport( blockName, 'anchor' );
